Split reusable app logic into library
